Changed to main loop, so the finalize functions wait for the threads to exit before the continue.
Fixed the way clients are created. Now the clients are created when
the engine is created, initialized when the engine is inited etc.

The engine now sets slow-sync if it has detected a problem in the last
sync.

added traces to some functions

changed the way error messages are treated and stored.

fixed the engine when dealing with timeouts


git-svn-id: http://svn.opensync.org/trunk@225 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e