Fix a few more issues in the python bindings.
The most important part is this:
-               if not hasattr(self, "__data"):
+               try:
+                       return self.__data
+               except AttributeError:
Python mangles names like __data to _OSyncChange__data, so the hasattr was 
always returning false. That meant that if you'd set the data from python, 
trying to read it back again returned garbage.

This change is possibly a bug:
+%include "cstring.i"
(...)
+       %cstring_input_binary(char *data, int datasize);
        void __set_data(char *data, int datasize) {
(...)
-               self.__set_data(data, len(data))
+               self.__set_data(data)
By getting SWIG to compute the size of data instead of using len(), we avoid 
any error that might occur if the elements in data are more than a byte long.

This change is really just a minor style issue:
-                       char *data = osync_change_get_data(self);
+                       void *data = osync_change_get_data(self);
                        int datasize = osync_change_get_datasize(self);
-                       char *copy = memcpy(malloc(datasize), data, datasize);
+                       void *copy = memcpy(malloc(datasize), data, datasize);
... use void * to make it clear that the data is an arbitrary buffer with a 
fixed length, and might not be nul-terminated.


git-svn-id: http://svn.opensync.org/trunk@1241 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e