Changed the layout of some tables in the Archive database!!!!!!!!!!! Those table change break groups .. you have to remove the archive.db file from the group directory or run "osyncdump $group --reset".
Every object type has now an own tbl_changes_$objtype
tbl_changelog_$objtype and tbl_archive_$objtype.

The initial reason was the wrong use of the merger which quierd only for
the memberid and not additional for the object type. So the merger
mixed up entries from e.g. event and contact entries.

This problem got disscussed on opensync-devel:
http://article.gmane.org/gmane.comp.misc.opensync.devel/1614

Those change made osync_archive_get_objtype() obsolate since the object
type for a table is fix. (API change!) Thats also the reason why
_osync_engine_receive_change() doesn't have to check if the object type
is known. Since the object type in a sink engine is the same all the
time.

Some testcases needed update ... because API changed and some requested
mapping entries changed the order in mappingtable and the winner of the conflict handling
needed to be twisted. (Related to changes in the mappingtable)



git-svn-id: http://svn.opensync.org/trunk@2168 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e