Made osync_member_find_objtype_sink() public which is needed by the previous conflict (getting support of certain conflict handlers).
FIXME: Tried to implement the detection if write or read sink function are
regiestered in the sinks of the plugin. And store the bool value if read
or write sink function are available or not. This _should_ be happen
after the initial discovery run... but doesn't work yet -> "read" will
always available - even if this funciton didn't get reported by the
discovery of the plugin.



git-svn-id: http://svn.opensync.org/trunk@2312 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e