Update the ListWrapper class, and use it wherever possible.
Rather than the previous prototype, where a new instance of the
wrapper class was instantiated on every access, this now overrides the
SWIG-generated constructors in a slightly messy way, but instantiates
the wrapper only once at construction time.

It doesn't make sense to define an __add__ method, as that is for
numeric addition. It may make sense to define an append() method, but
I've left that for later (keeping the list wrapper read-only has some
benefits IMO).

Also provide default NULL values for some parameters, where that makes
sense, and clean up a couple of other minor formatting nits I noticed.


git-svn-id: http://svn.opensync.org/trunk@2367 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e