Added further test facilities (aka. "dirty hacks") for testing purpose.
mock_get_error() needs to know the member number of the group - (or
something else to identifiy the member/plugin for certain tests.)

mock_get_error() allows to instrument certain "fake" errors, with
setting environmenta variables. This is used in engine-errors unittests.

To get mock_get_error() working, OSyncPluginInfo struct got extended
with the memberid (only when build with testing enabled!). This required
also the enhance the ipc message to add antoehr field in the initialize
message.


(This is a very ugly solution... but this is the only way without
 wasting the public API - OSyncClient* is acutally kind of independent
 of Memeber and GroupEnv - maybe there is a better way to identify the
 different members/plugins in the testcase.)



git-svn-id: http://svn.opensync.org/trunk@2990 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e