Add a Preferred element to the Formats one from plugin config.
This is propagated to the conversion path builder where it give more
weight to the given preferred format over the other with the same
conversion losses, objtype changes and conversion depth.

This is why I called it preferred, as it is not a mandatory format step,
only a hint for the conversion path builder.

The vertice idand neighbour_id are to fix the behaviour of the
conversion tree builder from :
->format0
<-format1.0 <- format 2.0 prefer format 2.0
->format 2.1 -> format 1.1 prefer format 1.1
etc

to
->format0
->format1.0->format 2.0 prefer format 1.0
-> format 1.1 -> format 2.1 prefer format 1.1
etc

The vertice id/neighbour_id change let/require some
plugins to set their wanted format. I tell requires because evo2-sync
for example tell it support vcard21 and vcard30 but believe that it will
have vcard30 always as input ... and with the fix to the conversion path
build process it will get vcard21. I ll fix it as soon as this patch or
anything close to it goes in. Plugin that tells they support more than
one format ought not to assume they will get one over the other blindly
anyway.



git-svn-id: http://svn.opensync.org/trunk@3659 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e