Introduce optional plugin sink function "connect_done". This plugin function gets called after connect() and before get_changes().
The function can be used for various purposes. Main reason for
introduction this sink function was to help synchronization protocols
which have to know if a Slow-Sync is required before completing the
entire connection-phase. So this function could be used to set the final
Slow-/Fast-Sync bit on the peer-protocol before finsihing the entire
connection-establishment process. This is something which got refered as
"late-Slow-Sync" in the past in various tickets.




git-svn-id: http://svn.opensync.org/trunk@4422 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e