Make use of "dummy" SinkEngine in OSyncObjEngine for improved mixed-objtype syncing.
If for a Proxy no suitable ObjTypeSink could be found, but the Proxy is
aware of ObjTypeSink "data", then a "dummy" SinkEngine got regiestered.
The "dummy" register will not call any Proxy Sink Function in
OSyncObjEngine, nor block any Engine Command, when not executed.

The only purpose of the "dummy" SinkEngine to assemble a list of
Mapping Entries which of mixed-objtypes, which need to get synced.

TODO: WRITE command of OSyncObjEngine is temp. calling the Proxy Sink
function of the dummy SinkEngine. This needs to get changed, that new
records/changes of "dummy" SinkEngine get written/committed by the
ObjTypeSink native OSyncObjEngine.


git-svn-id: http://svn.opensync.org/trunk@4455 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e