Avoid unexpected writes of changes which are equal on multi-member-sync-groups. Example scenario:
(member/entry)
Member 1         EntryA
Member 2         EntryA
Member 3

On a slow-sync, EntryA for Member 1 and 2 (both report the exact same
change!) get reported as ADDED. Expected result would be that it get
multiplied to only get written to Member 3.

This was not the case before. (EntryA for Member 2 or 1 got written as
modified before).

Another positive side-effect of this fix is on a conflict with a
multi-member-sync-group (more then two Members) where at least two
members report the exact change and another member reports a similar
change - since osync_mapping_engine_conflict() elimiates all same
changes, and just keeps one as reference, this allows to provide
conflict dialogs which just show the conflict changes of three different
member. The OSyncChange pointers of the same entries on a
multi-member-sync-groups are equal.

This got spotted by multisync unittest and testcase:
multisync_dual_new
multisync_trible_new



git-svn-id: http://svn.opensync.org/trunk@4467 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e