Introduce new engine command OSYNC_ENGINE_COMMAND_MULTIPLY. This also to make the "multipying" independent of the "writing" engine-commands. This solves also cosmetic issues on error conditions. (e.g. "all conflicts solved" on an aborted synchronization)
The purpose of this split is to abstract the multipying from the write.
This makes it later possible to transfer changes between different
OSyncObjEngines. This is related to the mixed-objtype synchronization
effort.

This allows also later to introduce a "write summary" before the actual
write. This is not only helpful for the user to decide to continue
syncing without dataloss, it's also very helfpul for debugging.

Needed to change various sync-tests case expected number of callback
function calls.

Testcase "sync_conflict_delay" got temporarliy disabled, the current
implementation doesn't allow aynchrounous conflict resolution in the
conflict callback. This will get fixed with the next
OSYNC_ENGINE_COMMAND-split: OSYNC_ENGINE_COMMAND_SOLVE_CONFLICTS or
something similar.

Further adoption for various error conditions were required.


git-svn-id: http://svn.opensync.org/trunk@4498 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e