Fix naming schema of osync_engine_num_objengines() by adding a 's'. osync_$module_num_$element_name_in_plural
Introduced opensync_engine_continue() for public API.
This is continuing paused/halted synchronization processes. This needs
to be done for instance inside the multiply callback.

Actived the multiply callback. It get called once the engine completed
the multiplying. If not multiplying callback is regestired the engine
continues with regular synchronization.



git-svn-id: http://svn.opensync.org/trunk@4897 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e