Introduced new OSyncArchive interface.
The new OSyncArchive doesn't require anylonger to get a file-location
specified. The database location is now handled internally and doesn't
require special handling by the plugin anylonger. This allows later to
easily add support further database-connectinos without any additional
porting of plugins.

Internally there is still a light dependency on sqlite3, which can be
fixed without breaking the public API of the OSyncArchive.

Porting instruction for plugins which using OSyncAnchor:

Most important changes:

osync_anchor_compare() doesn't return the comparsion result anylonger!
Instead it returns FALSE on an error when quering the database in the
background, TRUE otherwise. An additional parameter got added to
determine if the anchor value is the same.

 
-       if (!osync_anchor_compare(anchorpath, key, value))
-                osync_objtype_sink_set_slowsync(sink, TRUE);
+       if (!osync_anchor_compare(anchor, value, &anchormatch, &error))
+               goto error;
+       if (!anchormatch)
#                osync_objtype_sink_set_slowsync(sink, TRUE);

Also the first argument of osync_anchor_compare() and
osync_anchor_update() changed. Instead of the anchor-database file
location, an OSyncArchive pointer is required.

You can get the pointer by calling:

        OSyncObjTypeSink *sink = osync_plugin_info_get_sink(plugin_info);
+       OSyncAnchor *anchor = osync_objtype_sink_get_anchor(sink);

The key parameter from osync_anchor_compare() and osync_anchor_update()
got completely removed. This is not required anylonger since the
OSyncArchive is Object Type Sink specific. So for each sink you have a
different OSyncArchive object. OSyncArchive is also available for the
main-sink!

Example of osync_anchor_udpate() port in the sync_done() function:

-       const char *key = "foo";
-       osync_anchor_update(anchorpath, key, value);
+       OSyncAnchor *anchor = osync_objtype_sink_get_anchor(sink); 
+       if (!osync_anchor_update(anchor, value, &error))
+               goto error;
 


Last but not least. OSyncAnchor is stil optional! If your plugin
requries an OSyncArchive, it needs to be requested when initailizing the
OSyncObjTypeSinks in the plugin initialize function, by calling:

+               /* Request an anchor from the framework. */
+               osync_objtype_sink_enable_anchor(sink, TRUE);

Otherwise the call of osync_objtype_sink_get_anchor() will return NULL.
Reference counting of the anchor is done inside the engine.



git-svn-id: http://svn.opensync.org/trunk@5021 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e