Introduce support for determining "mixed objtypes".
This commit addresses following problem:

To create/initialize an OSyncObjEngine at least two member are required
which share the same ObjType. With mixed-objtype this this would not
work. Example:

          (objtype)
Member 1: event
Member 2: calendar


The engine would fail in initializing since it would not found a pair of
equal objtypes. This commit adds some logic to detect how many possible
ObjEngine could be initialized, and would not restrict them to at least
one. With the introduction with "dummy" OSyncSinkEngine, an
OSyncObjEngine can be created also only with one membr which has "real"
support of the ObjType.

The detection of this "mixed objtype" is based on format-conversion.
Format conversion allows allows in a very generic way to find relation
between different objtypes, due to their different formats. All possible
conversion pathes get created to determine if "event" and "calendar" can
synced together.



git-svn-id: http://svn.opensync.org/trunk@5237 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e