Fix testcase engine_sync_read_write. Regression got introduced by r5237 (mixed-objtype related modification)
test-fixture "group5" (from _create_group5) create two members with
different set of formats/objtypes. The second member had additionally
mockformat2/mockobjtype2 and mockformat3/mockobjtype3. Member 1 didn't.
So the engine failed syncing, since there were no further
peer-sinkengine which would support this anyway. Since this testcase
doesn't is intended to this this, the two additional format got dropped.
This still needs to be handled in the osync_engine_initialize()
function, by an error/fail. And not in the actual sync() call #992


git-svn-id: http://svn.opensync.org/trunk@5239 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e