Added "osync_bool slowsync" to the connect_done plugin sink function paramter list. To determine the current sync mode (slow-/fast-sync).
Also fixed a bug where connect_done() got called with the wrong sync
mode. Due to the sepeartion of Client and ClientProxy and the missing
update of the Sync Mode the connect_done() got called with fast-sync
even if the engine was supposed to do a slow-sync and would call
slow-ysnc with get_changes().

Added regression testcase inside mock-sync.




git-svn-id: http://svn.opensync.org/trunk@5422 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e