OSyncMappingEngine & OSyncMappingEntryEngine's don't get freed
When the engine is finalized, it doesn't look like the
OSyncMappingEngines or OSyncMappingEntryEngines are getting freed. I
checked by adding log messages to the *_unref() functions.

It looks like that an OSyncMappingEngines holds a reference on each
OSyncMappingEntryEngines and each OSyncMappingEntryEngines holds a
reference on the OSyncMappingEngines. A nice circle.

Original patch by marka
Sligthly adapted (coding style: // to /* */)

fixes #1122




git-svn-id: http://svn.opensync.org/trunk@5648 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e