Fix for mapping_engine_same_similar_conflict which was randomly failing.
Sometimes I see the mapping_engine_same_similar_conflict test fail on my
system. I tracked the issue down to the order in which the mock_sync
plugin reports changes in mock_report_dir. If entryA is reported before
entryB then the test passes. If entryB is reported first then the test
fails.

I have 2 patches. The first one modifies mock_sync to sort the changes
by file name before reporting them. It also adds a test which takes
advantage of that to reliably reproduce this issue.

The second patch tries to address the root cause of the issue. I'm not
completely familiar with the opensync internals so it may need some
work. It works for me and doesn't break any of the other tests.


Patch by marka

fixes #1123



git-svn-id: http://svn.opensync.org/trunk@5649 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e