Fix invalid comparsion results. Mixing up the OSyncChange pointers caused that the logical same entries got compared together. This was not obvious due to the cloning.
This bug was introduced by r5747

Overwriting pointer change1 too early with clone_change1 caused that the clone_demerge function
processed the logical change2, instead of change1. The twist of
demerging with capabililtiies of member 1, the change of member 2 and
vice-versa, is intended.


Report and most of the debugging was done by henrik




git-svn-id: http://svn.opensync.org/trunk@5869 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e