Fixed bug where member_id is not updated when new member is added
This patch sets the member_id in osync_group_add_member() in the same
place that the configdir is set, since that's when
osync_group_create_member_id() is called.

With this fix, it is possible to add a member to a group, and then
immediately list it, without reloading from disk.



git-svn-id: http://svn.opensync.org/trunk@5872 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e