API changes for osync_client_run_and_block()
Added parameter: OSyncError** which gets set on any error. Additionally
the return type changed to osync_bool. The return value is FALSE on any
error, on success TRUE.

New OSyncClient interfaces:
 * osync_client_{get,set}_plugin
 * osync_client_{get,set}_pipe_path

This changes allow to assemble plugins which run as an external process:
OSYNC_START_TYPE_EXTERNAL

Such plugins assemble OSyncPlugin* and OSyncClient* and need to set the
pipe path to the "pluginpipe" in the member direcotry.



git-svn-id: http://svn.opensync.org/trunk@5894 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e