Add testcase for #1207
Ie if a change is shared (members have "same" entries) , the conversion in 
prepare write will apply more than once to this change.
As the conversion path is cached, conversion will happens with a forced conversion
path on data already converted to another format.
Which can segfault and is anyway of no use and harmfull( change is not meant to be written).



git-svn-id: http://svn.opensync.org/trunk@6034 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e