Removed all unnecessary long longs from IDs, and gave IDs typedef names
The new ID types are:

        typedef int osync_memberid;
        typedef int osync_mappingid;
        typedef int osync_groupid;
        typedef long long int osync_archiveid;
        typedef long long int osync_messageid;

osync_archiveid is 64 bits because the ID is the row ID of the underlying
database.  This must match the database used.

Note that there is a bug in the implementation where the above 64 bit
IDs are crammed into a 32 bit pointer on 32 bit systems.  The
implementation should be fixed, not the size of the osync_archiveid.
Look for FIXME's in the code.

osync_messageid is 64 bits because 6 bytes are used for the time in
seconds and microseconds, and the lower 2 bytes are a random number,
to create a unique IPC message ID.

The other 3 IDs usually start at 1 or 0 and increment from there,
and int should be completely sufficient for them.



git-svn-id: http://svn.opensync.org/trunk@6051 53f5c7ee-bee3-0310-bbc5-ea0e15fffd5e