Revert "Added ${pc_top_builddir} to libopensync1.pc.in"
This reverts commit d1779ea4e1832f460d6a03fa01712ec44fa0c7ad

If the environment variable PKG_CONFIG_TOP_BUILD_DIR is not set,
then pkg-config returns a literal string of '$(top_builddir)' in
the cflags result.

On autoconf systems, this translates to '.' which breaks the build.

On cmake systems, this translatest to the source path, plus the
'$(top_builddir)' literal, which also breaks the build.

Basically, this is a broken idea.