Install python wrappers into system specific site-packages directory
With FindPythonLibs.cmake removed from the global cmake-modules tree,
the opensync-hacked PYTHON_VERSION was now missing.

Instead of going back to our own special version of FindPythonLibs.cmake,
this commit adds just the part we need.

Note that on Ubuntu and Debian, for python2.6, site-packages has been
renamed to dist-packages, so it seems better to ask python where the
whole directory is, instead of just asking for the version.

I'm not sure I've got this quite perfect.  Please email me with suggestions
if you have them.

See the following for background information:

https://lists.ubuntu.com/archives/ubuntu-devel/2009-February/027439.html
https://bugs.launchpad.net/ubuntu/+source/python2.6/+bug/362570
http://stackoverflow.com/questions/122327/how-do-i-find-the-location-of-my-python-site-packages-directory
http://blog.ianbicking.org/site-packages-considered-harmful.html