API: changed osync_capability_new() API
Changed behaviour and prototype of osync_capability_new(),
which now does not add itself to an OSyncCapabilitiesObjType
automatically, and must be unref() by the caller as usual
for _new() functions.

Note that this does not have much impact on usage, since
capabilities are usually accessed through _parse(), which
will do the OSyncCapabilitiesObjType add instead.

This commit updates the test code and the python code, such as
it is.