API: changed capability-related _parse() and _parse_and_add() functions
Before, some parse functions would return an object, and add it to an
internal list, and it was unclear as to whether the caller was supposed
to unref() the return value.

This commit changes the capability-related parse() calls to behave more
consistently.  All public and internal _parse() and _parse_and_add()
calls now return an object that you must _unref().  The _and_add()
versions add a ref'd copy to some internal list, which will be freed
automatically.