API: reorganized capabilities API to be more consistent and clear
The following API functions have been moved to internal, and are no longer
exposed.  The reason is that they are too cumbersome to use, now that they
are consistent in their _new() ref behaviour, and just by having them
exposed, it adds confusion for the user of the API.

	osync_capabilities_objtype_new
	osync_capabilities_objtype_ref
	osync_capabilities_objtype_unref
	osync_capability_new
	osync_capability_ref
	osync_capability_unref

To replace the functionality lost by the above, the following wrapper
functions have been added to the OSyncCapabilies level.  They behave
just like osync_capabilities_objtype_new and osync_capability_new
used to behave before these changes, but now the naming is clearer, and
it is impossible to ref or unref the returned pointers incorrectly.

	osync_capabilities_add_new_objtype
	osync_capabilities_add_new_capability