API: osync_xmlformat_assemble() now returns osync_strdup() buffer
osync_xmlformat_assemble() used to return a buffer allocated with
libxml2's memory functions, and therefore expected to be
freed with xmlFree().

This has been changed to return a buffer allocated with osync_strdup(),
so the caller can free it with osync_free() as expected.  Indeed most
of the code that I saw assumed this already.